requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-ratelimit==4.1.0
python-jose==3.5.0
langchain==0.1.0
langchain-ollama==0.1.0
Pillow==10.0.0
python-dotenv==1.0.0
python-magic==0.4.27
django-cleanup==8.1.0
cryptography==42.0.5
whitenoise==6.6.0
gunicorn==21.2.0 